server/requirements.txt
flask
torch
diffusers>=0.27.0
transformers>=4.40.0
accelerate
opencv-python
Pillow
numpy
